chore(common): create separate entrypoints for dsco common
