+++ frontend/packages/component-library/package.json
@@ -67,13 +67,13 @@
       "import": "./dist/esm/closeButton/index.mjs",
       "require": "./dist/cjs/closeButton/index.js"
     },
-    "./common": {
+    "./common/constants": {
       "types": {
-        "import": "./dist/esm/common/index.d.mts",
-        "require": "./dist/cjs/common/index.d.ts"
+        "import": "./dist/esm/common/constants/index.d.mts",
+        "require": "./dist/cjs/common/constants/index.d.ts"
       },
-      "import": "./dist/esm/common/index.mjs",
-      "require": "./dist/cjs/common/index.js"
+      "import": "./dist/esm/common/constants/index.mjs",
+      "require": "./dist/cjs/common/constants/index.js"
     },
     "./common/contexts": {
       "types": {
@@ -83,6 +83,14 @@
       "import": "./dist/esm/common/contexts/index.mjs",
       "require": "./dist/cjs/common/contexts/index.js"
     },
+    "./common/helpers": {
+      "types": {
+        "import": "./dist/esm/common/helpers/index.d.mts",
+        "require": "./dist/cjs/common/helpers/index.d.ts"
+      },
+      "import": "./dist/esm/common/helpers/index.mjs",
+      "require": "./dist/cjs/common/helpers/index.js"
+    },
     "./common/hooks": {
       "types": {
         "import": "./dist/esm/common/hooks/index.d.mts",
@@ -91,13 +99,13 @@
       "import": "./dist/esm/common/hooks/index.mjs",
       "require": "./dist/cjs/common/hooks/index.js"
     },
-    "./common/styles": {
+    "./common/types": {
       "types": {
-        "import": "./dist/esm/common/styles/index.d.mts",
-        "require": "./dist/cjs/common/styles/index.d.ts"
+        "import": "./dist/esm/common/types/index.d.mts",
+        "require": "./dist/cjs/common/types/index.d.ts"
       },
-      "import": "./dist/esm/common/styles/index.mjs",
-      "require": "./dist/cjs/common/styles/index.js"
+      "import": "./dist/esm/common/types/index.mjs",
+      "require": "./dist/cjs/common/types/index.js"
     },
     "./dialog": {
       "types": {
