+++ frontend/packages/component-library/src/common/helpers/index.ts
@@ -1,6 +1,6 @@
 import React, {AriaAttributes} from 'react';
 
-import {ComponentPlacementDirection} from './types';
+import {ComponentPlacementDirection} from '@/common/types';
 
 type PrimitiveValue =
   | string
