--- frontend/packages/component-library/src/common/index.ts
@@ -1,3 +1,0 @@
-export * from './types';
-export * as constants from './constants';
-export * as helpers from './helpers';
